backend/pyproject.toml
[project]
name = "autoipindia"
version = "0.1.0"
description = "Automated Indian Trademark Status Tracker"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "duckdb>=1.3.2",
    "google-genai>=1.26.0",
    "pandas>=2.3.1",
    "playwright>=1.53.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.28.0",
    "pydantic>=2.0.0",
    "duckdb-engine>=0.17.0",
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pytest>=7.4.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]
